requirements.txt
pyserial
docopt
